rewrite generate card-film from mocks
