+++ src/mock/card-film.js
@@ -16,7 +16,40 @@ import {
   NAMES
 } from "../const";
 
-import {generateComments} from "./comments";
+const AUTHORS = [
+  `Tim`,
+  `John`,
+  `Frank`
+];
+
+const COMMENTS = [
+  `Booo`,
+  `Wow!`,
+  `Greate!`,
+  `Fine!`,
+  `Bad..`
+];
+
+const EMOTIONS = [
+  `smile`,
+  `puke`,
+  `angry`,
+  `sleeping`
+];
+
+export const generateComments = () => {
+  const result = [];
+  const commentsCount = getRandomInteger(0, 5);
+  for (let i = 0; i < commentsCount; i++) {
+    result.push({
+      emotion: getRandomArrayItem(EMOTIONS),
+      author: getRandomArrayItem(AUTHORS),
+      text: getRandomArrayItem(COMMENTS),
+      commentDate: generateDate(),
+    });
+  }
+  return result;
+};
 
 const generateDescription = () => {
   const sentences = TEXT.split(`. `);
